msm: msm_dcvs: Add msm_dcvs manager
The msm_dcvs manager works with the DCVS algorithm in TrustZone by
taking regular idle pulse inputs and feeding back the frequency to
the CPU and GPU cores.

Core frequency is effected on the CPU by using a governor for the
CPUFreq framework and on the GPU using the KGSL driver.

Change-Id: I013c6070843f0ad19e0d2ea6087992b0d383cd44
Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>