msm_fb: hdmi: Add DVI sysfs node
For DVI mode, audio should not be routed. This routing happens
through HDMI daemon in usespace after confirming the mode from
HDMI core. This change gives HDMI daemon a way to determine if
the mode is DVI or HDMI, based on that HDMI Daemon can decide
on routing the audio.

Change-Id: I4afa746d6195b02046305a33a2b6a8b693510ba6
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>