gpu: ion: Remove protection of heap on first alloc
ION protects the heap when the first secure allocation
is done while unprotecting when the last secure allocation
is freed. In addition ION provides a separate API for protecting
and unprotecting a heap allowing users to better time the
protection of a heap.

The protection/unprotection done during first allocation and
last free is not needed anymore. Clients need better control
and thus use the separate API.

Remove protection of heap that is occuring when the
first secure allocation is done. Remove unprotecting of
the heap when the last secure allocation is freed.

Change-Id: I80254c3a79399cd4900a389dcb535d843d6dfa80
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>