gpu: ion: Add check for cached kernel mappings
An ION content protection heap should not allow
cached kernel space mapping for a heap that is
protected. Also, ION should not allow a heap that
has outstanding cached kernel space mappings to
be protected. These checks will ensure that the kernel
doesn't try to access memory that is locked down and
it cannot access.

Change-Id: I1f4a392d94ae8ae59c52978cfd0890e1a13fd39a
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>