msm: kgsl: A3XX GMEM save/restore was never executed
GMEM save/restore was never being executed for A3XX due to an
if() statement checking a flag that never gets set.  Remove said
if statement and set the flag so that GMEM save/restore gets run
during a context switch if requested from user space.

Additionally, fix a few minor issues with the gmem save/restore
sequence so when the application requests it, it will actually
work.

Change-Id: Ic0dedbad3f8e86b23bbd26051b771514e6518831
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>