msm: kgsl: Enable A3XX GPU hang detection
A3XX GPU hang detection can figure out if the core is
hung and interrupt us if it has.  However, the interrupt
seems to be a little unreliable so that isn't of use to
us, but the hang detection has a beneficial side effect
of spewing lots of useful information to debug registers
that can be used on a hang.  Enable the hang detection but
disable the hang interrupt.

Change-Id: Ic0dedbadc8492d388729c2eea9b9e1bdc8ac1013
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>