msm: kgsl: Disable ME, PFP split timeout and VFD interrupts.
The ME and PFP split timeout interrupts are fired when a wait for
idle does not get a response within the specified time limit, which
happens to be far too low for this revision of the hardware. The VFD
interrupt is harmless since we aways set the VFD range to the maximum
addressible size.  All these interrupts can be safely removed.

Change-Id: I32e714782a3c5a042b64360c1f6bcb7f8a9750db
Signed-off-by: Tarun Karra <tkarra@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>