msm: kgsl: Limit buffer dumping in the snapshot to the current context
In the vast majority of cases the only interesting indirect buffers (IBs)
 to dump in a shapshot are the ones that are in the current context
because they are the ones that probably contributed to the hang. Limit
the IBs that are dumped to the range between the start of context and
the RB read pointer. If the start of context can't be found, then dump
from the start of the ringbuffer dump to read pointer. Also, since
buffers can be reused (though typically not in the same context), use the
most recent parsed size when dumping the IB to make sure we get the right
buffer contents for the most recent usage of said buffer.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(cherry picked from commit e9e91bf3e498aeb58bc4e890c11e4f03e8ff8185)

Change-Id: I169ee434f8c234966e0d26edf5c29275bab91c74
Signed-off-by: Gopal G Goberu <ggober@codeaurora.org>
Signed-off-by: Ronny Schmatzler <schmatzler@googlemail.com>